requirements.txt
asgiref==3.5.0
astroid==2.11.2
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.4
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pylint==2.13.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.7
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
wrapt==1.14.0
